If ag sees its input stream has been redirected it searches in the input stream instead of files on disk. (This behavior can be overridden with an undocumented "search-files" command line option)
By default Node's ChildProcess.spawn method "connects" pipes to stdin, stdout, and stderr of the spawned child process.
On Windows (https://github.com/kjk/the_silver_searcher) the pipe connected to its stdin makes ag think it should search in the input stream instead of files -- to the user it effectively looks like atom-fuzzy-grep is never finding anything or is stuck.
On Linux everything works fine without this fix somehow, but it will not hurt anyway as there is no reason to redirect the child process' stdin because it is not used by atom-fuzzy-grep's code anyway.