Allow command line arguments in search
This commit allows extra command line arguments to be entered in the
search dialog. Some example use cases.

(1) You want to do a case sensitve search, but keep the insensitive defaults:

    -s COPYRIGHT

(2) You want to limit a search to a specific language:

    -t py -s COPYRIGHT

(3) You want to do multiple searches:

    -s -e Copyright -e COPYRIGHT

Use '--' if you need to do a search expression with spaces with a argument:

    -t py -- AUTHORS OR COPYRIGHTS